+++ .storybook/preview.js
@@ -24,6 +24,7 @@ const loaderFn = () => {
         require('../src/stories/Welcome.stories.mdx'),
         require('../src/components/Button/button.stories.mdx'),
         require('../src/components/Input/input.stories.mdx'),
+        require('../src/components/AutoComplete/autocomplete.stories.mdx'),
         require('../src/components/Alert/alert.stories.mdx'),
         require('../src/components/Menu/menu.stories.mdx'),
         require('../src/components/Menu/menuitem.stories.mdx'),
@@ -34,4 +35,3 @@ const loaderFn = () => {
     ]
 }
 configure(loaderFn, module);
-